requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-dotenv>=0.19.0

# Document processing
python-pptx>=0.6.21
PyPDF2>=2.0.0
PyMuPDF>=1.21.1
pytesseract>=0.3.8
Pillow>=8.3.1
opencv-python>=4.5.3
comtypes>=1.1.14

# Vector database
weaviate-client>=3.0.0

# ML and embeddings
torch>=1.9.0
transformers>=4.11.0
sentence-transformers>=2.0.0
scikit-learn>=0.24.2
nltk>=3.6.3
openai>=0.27.0

# Storage
boto3>=1.18.0

# Utilities
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.62.0 